+++ src/core/useNotifications.ts
@@ -1,16 +1,6 @@
 import { ref } from 'vue';
 
-export type NotificationType = 'success' | 'error' | 'info' | 'warning';
-export type NotificationPosition = 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right';
-
-export interface Notification {
-  id: number;
-  title?: string;
-  message: string;
-  type?: NotificationType;
-  duration?: number;
-  position?: NotificationPosition;
-}
+import type { NotificationType, NotificationPosition, Notification } from '../types/notifications'
 
 const notifications = ref<Notification[]>([]);
 let seed = 0;
