+++ src/plugin.ts
@@ -1,13 +1,9 @@
 import { notificationManager } from './core/NotificationManager';
 import { notificationStore } from './core/useNotifications';
 
-export function notify(options: {
-  title?: string;
-  message: string;
-  type?: 'success' | 'error' | 'info' | 'warning';
-  duration?: number;
-  position?: 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right';
-}) {
+import type { NotificationOptions } from './types/notifications'
+
+export function notify(options: NotificationOptions) {
   notificationManager.mount();
   notificationStore.add(options);
 }
